sessiondatapb: move SequenceCache to LocalOnlySessionData
This actually affects the results of my session migration test, so I
think we actually need to include it.

Release note: None
