+++ pkg/sql/sessiondata/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "internal.go",
         "search_path.go",
-        "sequence_cache.go",
         "sequence_state.go",
         "session_data.go",
     ],
