+++ pkg/sql/sessiondatapb/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     name = "sessiondatapb",
     srcs = [
         "local_only_session_data.go",
+        "sequence_cache.go",
         "session_data.go",
     ],
     embed = [":sessiondatapb_go_proto"],
